add --check option when saving the shrinkwrap file
if check is enabled we will verify if packages can be re-installed
reliably and warn if problems are found

There are 2 kind of problems that can be detected.
- version in component.json file is not the same as tagged version - in
  which case we just warn the user
- version in component.json is not tagged at all - in which case we set
  it to '*' in shrinkwrap file which causes master to be installed

In both cases the best course of action is to notify the component owner
and ask the components to be properly tagged.